requirements.txt
annotated-types
anyio
boto3
botocore
click
dnspython
exceptiongroup
fastapi
h11
httptools
idna
jmespath
pydantic
pydantic_core
python-dateutil
python-dotenv
PyYAML
s3transfer
six
sniffio
starlette
typing_extensions
urllib3
uvicorn
uvloop
watchfiles
websockets
opensearch-py
retrying
utility